% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Pickands}
\alias{Pickands}
\title{Pickands}
\usage{
Pickands(x, pn, alpha = 0.0027)
}
\arguments{
\item{x}{data}

\item{pn}{the nominal coverage probability}

\item{alpha}{false alarm rate}
}
\value{
A list containing the estimated tail index, control limits, the width of the control limits, sample size, and the nominal coverage probability.
\item{Tail_Index_Upper}{uppler tail index}
\item{Tail_Index_Lower}{lower tail index}
\item{UCL}{upper control limit}
\item{LCL}{lower control limit}
\item{Band.Pickands}{width of control limits}
\item{n}{sample size}
\item{pn}{the nominal coverage probability}
}
\description{
Pickands method
}
\examples{
data(ucisecom)
x1 = prepro(ucisecom$V2)[1:61]
n = length(x1)
Pickands(x1, pn=Pn_Pick(n))
}
