% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{Pickands}
\alias{Pickands}
\title{Pickands}
\usage{
Pickands(x, pn, alpha = 0.0027)
}
\arguments{
\item{x}{data}

\item{pn}{the nominal coverage probability}

\item{alpha}{false alarm rate}
}
\value{
A list containing the estimated tail index, control limits, and the width of the control limits.
\item{Tail_Index_Upper}{uppler tail index}
\item{Tail_Index_Lower}{lower tail index}
\item{UCL}{upper control limit}
\item{LCL}{lower control limit}
\item{Band.Moment}{width of control limits}
}
\description{
Pickands method
}
\examples{
data(ucisecom)
x = prepro(ucisecom$V2)
n = length(x)
Pickands(x, pn=Pn_Pick(n))
}
